testcontainers/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testcontainers-streamline"
version = "0.2.0"
description = "Testcontainers module for Streamline - The Redis of Streaming"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Jose David Baena", email = "josedab@users.noreply.github.com"}
]
keywords = ["streamline", "kafka", "testcontainers", "testing", "docker"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.9"
dependencies = [
    "testcontainers>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "kafka-python>=2.0.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/streamlinelabs/streamline-python-sdk"
Documentation = "https://github.com/streamlinelabs/streamline-python-sdk/tree/main/sdks/testcontainers-python"
Repository = "https://github.com/streamlinelabs/streamline-python-sdk"
Issues = "https://github.com/streamlinelabs/streamline-python-sdk/issues"

[tool.setuptools.packages.find]
include = ["streamline_testcontainers*"]
